centralize validation

+++ src/main/java/org/jenkinsci/plugins/fodupload/FodUploaderPlugin.java
@@ -12,6 +12,7 @@ import jenkins.tasks.SimpleBuildStep;
 import org.jenkinsci.plugins.fodupload.models.JobModel;
 import org.kohsuke.stapler.DataBoundConstructor;
 
+import javax.annotation.Nonnull;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -38,7 +39,6 @@ public class FodUploaderPlugin extends Recorder implements SimpleBuildStep {
 
     private static final ThreadLocal<TaskListener> taskListener = new ThreadLocal<>();
 
-    private static FodApi api;
     private static JobModel model;
 
     // Fields in config.jelly must match the parameter names in the "DataBoundConstructor"
@@ -61,12 +61,13 @@ public class FodUploaderPlugin extends Recorder implements SimpleBuildStep {
 
     // logic run during a build
     @Override
-    public void perform(Run<?, ?> build, FilePath workspace, Launcher launcher, TaskListener listener) {
+    public void perform(@Nonnull Run<?, ?> build, @Nonnull FilePath workspace,
+                        @Nonnull Launcher launcher, @Nonnull TaskListener listener) {
         try {
             final PrintStream logger = listener.getLogger();
             taskListener.set(listener);
             // Load api settings
-            api = getDescriptor().createFodApi();
+            FodApi api = getDescriptor().createFodApi();
 
             if (api == null) {
                 logger.println("Error: Failed to Authenticate with Fortify API.");
@@ -74,35 +75,33 @@ public class FodUploaderPlugin extends Recorder implements SimpleBuildStep {
             } else {
                 api.authenticate();
 
-                logger.println("Starting FoD Upload.");
+                // Add all validation here
+                if (model.validate(logger)) {
+                    logger.println("Starting FoD Upload.");
 
-                if (model.getBsiUrl().getAssessmentTypeId() == 0) {
-                    logger.println("Assessment Type is empty.");
-                    build.setResult(Result.FAILURE);
-                }
-
-                // zips the file in a temporary location
-                File payload = CreateZipFile(workspace);
-                if (payload.length() == 0) {
-                    logger.println("Source is empty for given Technology Stack and Language Level.");
-                    build.setResult(Result.FAILURE);
-                }
-                logger.println(model.getBsiUrl().ORIGINAL_VALUE);
-
-                model.setPayload(payload);
-                boolean success = api.getStaticScanController().startStaticScan(model);
-                boolean deleted = payload.delete();
-                if (success && deleted) {
-                    logger.println("Scan Uploaded Successfully.");
-                    if (getDescriptor().getDoPollFortify() && model.getPollingInterval() > 0) {
-                        PollStatus /*Amy*/poller = new PollStatus(api, model);
-                        success = poller.releaseStatus(model.getBsiUrl().getProjectVersionId());
+                    // zips the file in a temporary location
+                    File payload = CreateZipFile(workspace);
+                    if (payload.length() == 0) {
+                        logger.println("Source is empty for given Technology Stack and Language Level.");
+                        build.setResult(Result.FAILURE);
                     }
-                }
 
-                // Success could be true then set to false from polling.
-                api.retireToken();
-                build.setResult(success ? Result.SUCCESS : Result.UNSTABLE);
+                    model.setPayload(payload);
+                    boolean success = api.getStaticScanController().startStaticScan(model);
+                    boolean deleted = payload.delete();
+                    if (success && deleted) {
+                        logger.println("Scan Uploaded Successfully.");
+                        if (getDescriptor().getDoPollFortify() && model.getPollingInterval() > 0) {
+                            PollStatus /*Amy*/poller = new PollStatus(api, model);
+                            success = poller.releaseStatus(model.getBsiUrl().getProjectVersionId());
+                        }
+                    }
+
+                    // Success could be true then set to false from polling.
+                    api.retireToken();
+                    build.setResult(success ? Result.SUCCESS : Result.UNSTABLE);
+                }
+                build.setResult(Result.FAILURE);
             }
         } catch (IOException e) {
             e.printStackTrace();
@@ -142,37 +141,39 @@ public class FodUploaderPlugin extends Recorder implements SimpleBuildStep {
     // NOTE: The following Getters are used to return saved values in the config.jelly. Intellij
     // marks them unused, but they actually are used.
     // These getters are also named in the following format: Get<JellyField>.
+    @SuppressWarnings("unused")
     public String getBsiUrl() {
         return model.getBsiUrl().ORIGINAL_VALUE;
     }
+    @SuppressWarnings("unused")
     public boolean getRunOpenSourceAnalysis() {
         return model.isRunOpenSourceAnalysis();
     }
-
+    @SuppressWarnings("unused")
     public boolean getIsExpressScan() {
         return model.isExpressScan();
     }
-
+    @SuppressWarnings("unused")
     public boolean getIsExpressAudit() {
         return model.isExpressAudit();
     }
-
+    @SuppressWarnings("unused")
     public boolean getDoPrettyLogOutput() {
         return model.isDoPrettyLogOutput();
     }
-
+    @SuppressWarnings("unused")
     public boolean getIncludeAllFiles() {
         return model.isIncludeAllFiles();
     }
-
+    @SuppressWarnings("unused")
     public boolean getExcludeThirdParty() {
         return model.isExcludeThirdParty();
     }
-
+    @SuppressWarnings("unused")
     public boolean getIsRemediationScan() {
         return model.isRemediationScan();
     }
-
+    @SuppressWarnings("unused")
     public int getPollingInterval() {
         return model.getPollingInterval();
     }
